--- main/Checkbox_/index.tsx
@@ -1,2 +1,0 @@
-export { default as Checkbox_ } from './Checkbox_';
-export { default as CheckboxGroup_ } from './CheckboxGroup_';
