+++ main/__tests__/Checkbox.test.tsx
@@ -1,27 +1,18 @@
 import * as React from 'react';
+import { Checkbox } from '../Checkbox';
 
-import { CheckboxGroup } from '../Checkbox';
+import {
+  render,
+} from '@testing-library/react-native';
 
-import renderer from 'react-test-renderer';
+// Note: test renderer must be required after react-native.
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-let props: any;
-let component: React.ReactElement;
-
-const createTestProps = (obj: Record<string, unknown>): Record<string, unknown> => ({
-  ...obj,
-});
-
-describe('[CheckboxGroup] render', () => {
-  beforeEach(() => {
-    props = createTestProps({});
-    component = <CheckboxGroup {...props} />;
-  });
-  it('renders without crashing', () => {
-    const rendered: renderer.ReactTestRendererJSON | null = renderer
-      .create(component)
-      .toJSON();
+describe('[CheckBox]', () => {
+  const defaultPros = {
+    label: 'testProps',
+  };
+  it('render without crashing', () => {
+    const rendered = render(<Checkbox {...defaultPros}/>);
     expect(rendered).toMatchSnapshot();
-    expect(rendered).toBeTruthy();
   });
 });
