+++ stories/dooboo-ui/Checkbox.stories.tsx
@@ -1,7 +1,8 @@
-import React, { ReactElement } from 'react';
+import React, { ReactElement, useState } from 'react';
+import { Text } from 'react-native';
 
 // eslint-disable-next-line sort-imports
-import { CheckboxGroup, CheckboxItem } from '../../main';
+import { Checkbox, CheckboxGroup } from '../../main';
 import { ContainerDeco } from '../../storybook/decorators';
 
 import { storiesOf } from '@storybook/react-native';
@@ -21,80 +22,125 @@ const Container = styled.View`
   flex-direction: column;
 `;
 
-const Selected = styled.Text`
+const Title = styled.Text`
   padding: 20px;
-  font-size: 13px;
+  font-size: 20px;
   font-weight: bold;
 `;
 
-const defaultCheckboxData: CheckboxItem[] = [
-  {
-    label: 'first',
-    value: 1,
-    selected: true,
-  },
-  {
-    label: 'second',
-    value: 2,
-    selected: false,
-  },
-  {
-    label: 'third',
-    value: 3,
-    selected: false,
-  },
-];
+const Seperator = styled.View`
+ border-bottom-color: black;
+ border-bottom-width: 1;
+`;
 
-const disabledCheckboxData: CheckboxItem[] = [
-  {
-    label: 'first',
-    value: 1,
-    selected: true,
-  },
-  {
-    label: 'second',
-    value: 2,
-    selected: false,
-  },
-  {
-    label: 'checked disabled value',
-    value: 3,
-    selected: true,
-    disabled: true,
-  },
-  {
-    label: 'unchecked disabled value',
-    value: 4,
-    selected: false,
-    disabled: true,
-  },
-];
-
-function Default(): React.ReactElement {
+function DefaultCheckbox(): React.ReactElement {
+  const customStyle = {
+    labelLeft: true,
+  };
   return (
     <ScrollContainer>
+
       <Container>
+        <Title>default</Title>
+        <Checkbox label="defaultChecked" defaultChecked ></Checkbox>
 
-        <Selected>Selected: {1}</Selected>
+      </Container>
 
-        <CheckboxGroup
-          boxSize={20}
-          boxColor={'orange'}
-          labelSize={20}
-          labelColor={'#000000'}
-          items={defaultCheckboxData}
+      <Container>
+        <Title>disabled</Title>
+        <Checkbox label="disabled" disabled />
+      </Container>
+
+      <Container>
+        <Title>indeterminate</Title>
+        <Checkbox label="indeterminate" indeterminate />
+      </Container>
+
+      <Container>
+        <Title>labelLeft</Title>
+        <Checkbox label="labelLeft" customStyle={customStyle} />
+      </Container>
+
+    </ScrollContainer>
+  );
+}
+
+function DefaultCheckboxGroup(): React.ReactElement {
+  const plainOptions = ['Apple', 'Pear', 'Orange'];
+  const options = [
+    { label: 'Apple', value: 'Apple' },
+    { label: 'Pear', value: 'Pear' },
+    { label: 'Orange', value: 'Orange' },
+    { label: 'Mango', value: 'Mango' },
+  ];
+  const optionsWithDisabled = [
+    { label: 'Apple', value: 'Apple' },
+    { label: 'Pear', value: 'Pear' },
+    { label: 'Orange', value: 'Orange', disabled: true },
+  ];
+
+  const onChange = (checkedValues): void => {
+    console.log('checked = ', checkedValues);
+  };
+  return (
+    <ScrollContainer>
+
+      <Container>
+        <Title>plainOption</Title>
+        <CheckboxGroup options={plainOptions} onChange={onChange} />
+      </Container>
+      <Container>
+        <Title>Default</Title>
+        <CheckboxGroup options={options} onChange={onChange} />
+      </Container>
+
+      <Container>
+        <Title>option with disabled</Title>
+        <CheckboxGroup options={optionsWithDisabled} onChange={onChange} />
+      </Container>
+
+      <Container>
+        <Title>with disabled props</Title>
+        <CheckboxGroup options={options} disabled onChange={onChange} />
+      </Container>
+
+    </ScrollContainer>
+  );
+}
+
+function CheckAll(): React.ReactElement {
+  const plainOptions = ['Apple', 'Pear', 'Orange'];
+  const defaultCheckedList = ['Apple', 'Orange'];
+
+  const [checkedList, setCheckedList] = useState(defaultCheckedList);
+  const [indeterminate, setIndeterminate] = useState(defaultCheckedList.length > 0);
+  const [checkAll, setCheckAll] = useState(false);
+
+  const onChange = (checkedList): void => {
+    setCheckedList(checkedList);
+    setIndeterminate(!!checkedList.length && checkedList.length < plainOptions.length);
+    setCheckAll(checkedList.length === plainOptions.length);
+  };
+
+  const onCheckAllChange = (e) : void => {
+    setCheckedList(e.checked ? plainOptions : []);
+    setIndeterminate(false);
+    setCheckAll(e.checked);
+  };
+
+  return (
+    <ScrollContainer>
+
+      <Container>
+        <Title>Check All</Title>
+        <Checkbox
+          label={'Check All'}
+          indeterminate={indeterminate}
+          onChange={onCheckAllChange}
+          checked={checkAll}
         />
-
-        <Selected>checkbox disabled</Selected>
-
-        <CheckboxGroup
-          boxSize={20}
-          boxColor={'orange'}
-          labelSize={20}
-          labelColor={'#000000'}
-          items={disabledCheckboxData}
-        />
-
+        <Seperator/>
+        <CheckboxGroup values={checkedList} options={plainOptions} onChange={onChange} />
       </Container>
     </ScrollContainer>
   );
@@ -104,15 +150,31 @@ export default {
   title: 'Checkbox',
 };
 
-export const toStorybook = (): ReactElement => <Default />;
+/**
+ * Below are stories for WEB
+ */
+
+export const DefaultCheckbox_ = (): ReactElement => <DefaultCheckbox />;
+export const DefaultCheckboxGroup_ = (): ReactElement => <DefaultCheckboxGroup/>;
+export const ReactElement_ = (): ReactElement => <CheckAll/>;
 
 /**
  * Below are stories for app
  */
 storiesOf('Checkbox', module)
   .addDecorator(ContainerDeco)
-  .add('default', () => (
+  .add('Checkbox', () => (
     <>
-      <Default />
+      <DefaultCheckbox />
+    </>
+  ))
+  .add('CheckboxGroup', () => (
+    <>
+      <DefaultCheckboxGroup />
+    </>
+  ))
+  .add('CheckAll', () => (
+    <>
+      <CheckAll />
     </>
   ));
