+++ stories/index.ts
@@ -10,7 +10,6 @@ import './dooboo-ui/SwitchToggle.stories';
 import './dooboo-ui/Slider.stories';
 import './dooboo-ui/Rating.stories';
 import './dooboo-ui/Checkbox.stories';
-import './dooboo-ui/Checkbox_.stories';
 import './packages/Charts.stories';
 import './packages/Snackbar.stories';
 import './packages/TinderCard.stories';
